+++ R/few.R
@@ -81,3 +81,50 @@ theme_few <- function(base_size = 12, base_family="") {
               strip.background = element_rect(fill = "white", colour = NA)
             )
 }
+
+#' Shape palette from "Show Me the Numbers" (discrete)
+#'
+#' Shape palette from Stephen Few's, "Show Me the Numbers".
+#' It consists of five shapes: circle, square, triangle, plus, times.
+#'
+#' @references Few, S. (2012)
+#'   \emph{Show Me the Numbers: Designing Tables and Graphs to Enlighten},
+#'   Analytics Press, p. 208.
+#'
+#' @export
+few_shape_pal <- function() {
+  max_n <- 5
+  f <- function(n) {
+    msg <- paste("The shape palette can deal with a maximum of ", max_n,
+                 "discrete values;",
+                 "you have ", n, ".",
+                 " Consider specifying shapes manually if you ",
+                 "must have them.", sep = "")
+    if (n > max_n) {
+      stop(msg, call. = FALSE)
+    }
+    # circle, square, triangle, plus, cross
+    c(1, 0, 2, 3, 4)[seq_len(n)]
+  }
+  attr(f, "max_n") <- max_n
+  f
+}
+
+#' Scales for shapes from "Show Me the Numbers"
+#'
+#' \code{scale_shape_few} maps discrete variables to five easily
+#' discernible shapes. It is based on the shape palette suggested in
+#' Few (2012).
+#'
+#' @param ... Common \code{discrete_scale} parameters. See
+#'   \code{\link[ggplot2]{discrete_scale}} for more details.
+#'
+#' @references Few, S. (2012)
+#'   \emph{Show Me the Numbers: Designing Tables and Graphs to Enlighten},
+#'   Analytics Press, p. 208.
+#' @seealso \code{\link{scale_shape_few}} for the shape palette that this
+#'   scale uses.
+#' @export
+scale_shape_few <- function(...) {
+  discrete_scale("shape", "few", few_shape_pal(), ...)
+}